batch: Move ownership merge helpers
Ownership metadata still housed absence signatures and deletion
deduplication, so callers depended on the data model for merge behavior.

The new ownership_merging module owns merge signatures, baseline-reference
selection, and replacement-unit remapping for merged deletion indexes.

This commit moves merge_batch_ownership and its support code out of
ownership metadata.

Callers now import merge_batch_ownership from the merge boundary while
ownership.py keeps model serialization and acquisition paths.
